requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
python-dotenv
sentence-transformers
google-generativeai